Fix DLSS for #230
Fixes DLSS by calculating a scale value using Screen.height divided by camera.scaledPixelHeight. Not sure if this is the best method to accomplish this as I am kind of new to C# but local testing shows it works and places all markers in the right spot at all DLSS quality settings. Additionally works with supersampling and downsampling settings.