Unlimited Ammo - First attempt
Functional except for revolvers and single shot weapons (MP-18, Rhino, maybe the Rsh?).  Duplicates ammo if less than full magazine due to the hack of just looking at all shots in the game same as the wallshoot. The correct method is to patch BallisticsCalculator.Shoot with Harmony but my understanding of how to do that in CSharp is minimal at best.