Wallhack Changes
Separates infoColor from color and borderColor. Changes boxes to use borderColor not color. Adds basic Shootable detection and configurable colors, ideally we would enumerate through each part of the enemy we are looking to check if the enemy could be hit but currently the check is only against the Head.

The separate colors and the shootable colors does cause an issue where the options exceed the size of the gui window and doesn't resize. I couldn't figure out where to fix this.

This addresses some of the requests in #299.