Added better shootability detection
Reworked Players.cs to perform an iterative check against enemy bones to determine if they are possibly shootable. We perform this check in a foreach loop for speed and break once we hit one as we don't draw the bones until later. In the event that none are hit we skip further checks for color and render in the base color of that enemy.  In Bones.cs two new lists are created for performance reasons as well as a different RenderBones() function that supports two colors. The first list created HitConnections[] is used for performing hit detection of major bones. The second list FingerConnections[] are then quickly drawn in not shootable color for performance reasons as doing an extra 60 visibility checks kind of tanks fps.