Fixed skeleton alignment and scope clamping
Works well when enemies are at distance. There is weirdness when target is close to player or if target is at the edges of the scope but I don't think this can be mitigated due to how we grab the position. Bones are now properly clamped to the edge of the scope circle.