Radar position fixes and Radar Compass
When drawing the full screen map we use WorldPointToScreenPoint which incorporates the DLSS scaling, for the minimap we use WorldToScreenPoint and subtract the screen height as for some reason this isn't affected by DLSS???. Reordered some variables to prevent having to add a second function within BaseMapToggleFeature. Also streamlined radarSize instead of sizeX and sizeY as we don't need both for a square.

Added in compass function that rotates with map view, some reason it inherits the enemy colors?