TrackWishlist in Loot feature
Removed loadwl command as we have now incorporated it into an option under the Loot feature. If the option is on we assign the Wishlist variable to the wishlist class G2903 or null as appropriate. Then when performing checks in TryAddRecordIfTracked() we shortcut if the item was found in the tracklist or if the Wishlist is null.